Geocoding/requirements.txt
requests==2.25.0
pytest==7.4.0
allure-pytest==2.13.2
